Moved Posture Perfection code to Examples Directory
