+++ .gitignore
@@ -5,6 +5,12 @@ CMakeFiles/*
 Makefile
 install_manifest.txt
 
+#PosturePerfection
+Examples/PosturePefection/CMakeCache.txt
+Examples/PosturePerfection/cmake_install.cmake
+Examples/PosturePerfection/CMakeFiles/*
+Examples/PosturePerfection/Makefile
+
 #Binaries
 libRemoteNotify.a
 libRemoteNotifyBroadcast.a
@@ -13,4 +19,5 @@ PosturePerfectionReceiver
 PosturePerfectionTestBroadcast
 #Vim
 .swp
-
+#VSCODE
+.vscode/
